add mercado pago integration

+++ packages/Webkul/MercadoPago/src/Database/Migrations/2024_01_01_000001_create_mercadopago_webhook_events_table.php
@@ -23,7 +23,7 @@ return new class extends Migration
             $table->string('ip_address')->nullable()->index();
             $table->text('user_agent')->nullable();
             $table->enum('status', ['received', 'processed', 'failed', 'error'])->default('received')->index();
-            $table->unsignedBigInteger('order_id')->nullable()->index();
+            $table->unsignedInteger('order_id')->nullable()->index();
             $table->string('payment_id')->nullable()->index();
             $table->timestamp('received_at')->index();
             $table->timestamp('processed_at')->nullable()->index();
